refactor to multiple components
